pyproject.toml
[project]
name = "ai-create-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-cpp-python>=0.3.16",
    "llama-index>=0.14.8",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-llama-cpp>=0.5.1",
    "llama-index-llms-ollama>=0.9.0",
    "llama-index-vector-stores-faiss>=0.5.1",
]
